Move some documentation utilities to utils.
These utilities were unavailable to all ARMI plugins because they were
not importable from doc/conf. This moved them to an importable place so
they can be used from elsewhere.
